Externalized the daemon thread factory
